--- example_default/index.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>ReactiveForms Default RendererSet</title>
-  <base href="/">
-
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <script src="assets/native-shim.js"></script>
-  <link rel="stylesheet" href="assets/example.css">
-  <link rel="stylesheet" href="assets/example.dark.css">
-  <!--<link rel="stylesheet" href="assets/example.labelfixed.css">-->
-
-  <script src="assets/bundle_default.js"></script>
-</head>
-<body>
-<div id="root">
-  <div id="toolbar">
-    <div id="examples"></div>
-    <div id="theme"></div>
-  </div>
-  <div id="view"></div>
-</div>
-</body>
-<script>
-  exampleDivId = 'examples';
-  viewDivId = 'view';
-</script>
-</html>
