+++ packages/materialized/src/complex/index.ts
@@ -1,0 +1,1 @@
+export * from './array-renderer';
