+++ packages/materialized/src/controls/index.ts
@@ -1,0 +1,2 @@
+export * from './material-boolean.control';
+export * from './material-input.control';
