+++ packages/materialized/src/layouts/index.ts
@@ -1,3 +1,3 @@
 export * from './group.layout';
-export * from './horizontal.layout';
+export * from './material-horizontal.layout';
 export * from './vertical.layout';
