--- test/renderers/date.field.test.tsx
@@ -1,414 +1,0 @@
-import { JSX } from '../../src/renderers/JSX';
-import test from 'ava';
-import DateField, { dateFieldTester } from '../../src/renderers/fields/date.field';
-import { ControlElement, HorizontalLayout } from '../../src/models/uischema';
-import { HorizontalLayoutRenderer } from '../../src/renderers/layouts/horizontal.layout';
-import { JsonForms } from '../../src/core';
-import { getData } from '../../src/reducers/index';
-import { update, validate } from '../../src/actions';
-import { JsonSchema } from '../../src/models/jsonSchema';
-import { initJsonFormsStore } from '../helpers/setup';
-import {
-  change,
-  findRenderedDOMElementWithClass,
-  findRenderedDOMElementWithTag,
-  renderIntoDocument,
-  scryRenderedDOMElementsWithTag
-} from '../helpers/binding';
-import { Provider } from '../../src/common/binding';
-
-test.before(() => {
-  JsonForms.stylingRegistry.registerMany([
-    {
-      name: 'control',
-      classNames: ['control']
-    },
-    {
-      name: 'control.validation',
-      classNames: ['validation']
-    }
-  ]);
-});
-
-test.beforeEach(t => {
-  t.context.data = { 'foo': '1980-04-04' };
-  t.context.schema = {
-    type: 'object',
-    properties: {
-      foo: {
-        type: 'string',
-        format: 'date'
-      },
-    },
-  };
-  t.context.uischema = {
-    type: 'Control',
-    scope: {
-      $ref: '#/properties/foo',
-    },
-  };
-});
-// TODO: don't understand this test case
-test.skip('autofocus on first element', t => {
-    const schema: JsonSchema = {
-        type: 'object',
-        properties: {
-            firstDate: { type: 'string', format: 'date' },
-            secondDate: { type: 'string', format: 'date' }
-        }
-    };
-    const firstControlElement: ControlElement = {
-        type: 'Control',
-        scope: {
-            $ref: '#/properties/firstDate'
-        },
-        options: {
-            focus: true
-        }
-    };
-    const secondControlElement: ControlElement = {
-        type: 'Control',
-        scope: {
-            $ref: '#/properties/secondDate'
-        },
-        options: {
-            focus: true
-        }
-    };
-    const uischema: HorizontalLayout = {
-        type: 'HorizontalLayout',
-        elements: [
-            firstControlElement,
-            secondControlElement
-        ]
-    };
-    const data = {
-        'firstDate': '1980-04-04',
-        'secondDate': '1980-04-04'
-    };
-    const store = initJsonFormsStore(
-        data,
-        schema,
-        uischema
-    );
-    const tree = renderIntoDocument(
-        <Provider store={store}>
-            <HorizontalLayoutRenderer schema={schema}
-                                      uischema={uischema}
-            />
-        </Provider>
-    );
-    const inputs = scryRenderedDOMElementsWithTag(tree, 'input');
-    t.not(document.activeElement, inputs[0]);
-    t.is(document.activeElement, inputs[1]);
-});
-
-test('autofocus active', t => {
-    const uischema: ControlElement = {
-        type: 'Control',
-        scope: {
-            $ref: '#/properties/foo'
-        },
-        options: {
-            focus: true
-        }
-    };
-    const store = initJsonFormsStore(t.context.data, t.context.schema, uischema);
-    const tree = renderIntoDocument(
-        <Provider store={store}>
-            <DateField schema={t.context.schema}
-                         uischema={uischema}
-            />
-        </Provider>
-    );
-    const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-    t.is(document.activeElement, input);
-});
-
-test('autofocus inactive', t => {
-    const uischema: ControlElement = {
-        type: 'Control',
-        scope: {
-            $ref: '#/properties/foo'
-        },
-        options: {
-            focus: false
-        }
-    };
-    const store = initJsonFormsStore(t.context.data, t.context.schema, uischema);
-    const tree = renderIntoDocument(
-        <Provider store={store}>
-            <DateField schema={t.context.schema}
-                            uischema={uischema}
-            />
-        </Provider>
-    );
-    const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-    t.false(input.autofocus);
-});
-
-test('autofocus inactive by default', t => {
-    const uischema: ControlElement = {
-        type: 'Control',
-        scope: {
-            $ref: '#/properties/foo'
-        }
-    };
-    const store = initJsonFormsStore(t.context.data, t.context.schema, uischema);
-    const tree = renderIntoDocument(
-        <Provider store={store}>
-            <DateField schema={t.context.schema}
-                            uischema={uischema}
-            />
-        </Provider>
-    );
-    const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-    t.false(input.autofocus);
-});
-
-test('tester', t => {
-  t.is(dateFieldTester(undefined, undefined), -1);
-  t.is(dateFieldTester(null, undefined), -1);
-  t.is(dateFieldTester({ type: 'Foo' }, undefined), -1);
-  t.is(dateFieldTester({ type: 'Control' }, undefined), -1);
-});
-
-test('tester with wrong prop type', t => {
-  t.is(
-      dateFieldTester(
-          t.context.uischmea,
-          {
-            type: 'object',
-            properties: {
-              foo: { type: 'string' },
-            },
-          },
-      ),
-      -1,
-  );
-});
-
-test('tester with wrong prop type, but sibling has correct one', t => {
-  t.is(
-      dateFieldTester(
-          t.context.uischema,
-          {
-            type: 'object',
-            properties: {
-              foo: { type: 'string' },
-              bar: {
-                type: 'string',
-                format: 'date',
-              },
-            },
-          },
-      ),
-      -1,
-  );
-});
-
-test('tester with correct prop type', t => {
-  t.is(
-    dateFieldTester(
-      t.context.uischema,
-      {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-            format: 'date',
-          },
-        },
-      },
-    ),
-    2,
-  );
-});
-
-test('render', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  t.is(input.type, 'date');
-  t.is(input.value, '1980-04-04');
-});
-
-test('update via event', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  input.value = '1961-04-12';
-  change(input);
-  setTimeout(() => t.is(getData(store.getState()).foo, '1961-04-12'), 100);
-});
-
-test('update via action', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  store.dispatch(update('foo', () => '1961-04-12'));
-  setTimeout(() => t.is(input.value, '1961-04-12'), 100);
-});
-
-test('update with null value', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  store.dispatch(update('foo', () => null));
-  t.is(input.value, '');
-});
-
-test('update with undefined value', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  store.dispatch(update('foo', () => undefined));
-  t.is(input.value, '');
-});
-
-test('update with wrong ref', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  store.dispatch(
-    update(
-      'bar',
-      () => 'Bar'
-  ));
-  t.is(input.value, '1980-04-04');
-});
-
-test('update with null ref', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  store.dispatch(
-    update(
-      null,
-      () => '1961-04-12'
-    ));
-  t.is(input.value, '1980-04-04');
-});
-
-test('update with undefined ref', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                   uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  store.dispatch(
-    update(
-      undefined,
-      () => '1961-04-12'
-    ));
-  t.is(input.value, '1980-04-04');
-});
-
-test('hide', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                      uischema={t.context.uischema}
-                      visible={false}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  t.true(input.hidden);
-});
-
-test('show by default', t => {
-  const store = initJsonFormsStore(
-    t.context.data,
-    t.context.schema,
-    t.context.uischema
-  );
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                      uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  t.false(input.hidden);
-});
-
-test('disable', t => {
-  const store = initJsonFormsStore(
-    t.context.data,
-    t.context.schema,
-    t.context.uischema
-  );
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                      uischema={t.context.uischema}
-                      enabled={false}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  t.true(input.disabled);
-});
-
-test('enabled by default', t => {
-  const store = initJsonFormsStore(t.context.data, t.context.schema, t.context.uischema);
-  const tree = renderIntoDocument(
-    <Provider store={store}>
-      <DateField schema={t.context.schema}
-                      uischema={t.context.uischema}
-      />
-    </Provider>
-  );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
-  t.false(input.disabled);
-});
